cleanup + reorganize
