+++ codex-rs/tui/src/app_event.rs
@@ -182,9 +182,6 @@ pub(crate) enum AppEvent {
         category: FeedbackCategory,
     },
 
-    /// Clear a pending external editor request if no draw occurred in time.
-    ExternalEditorRequestTimeout,
-
     /// Launch the external editor after a normal draw has completed.
     LaunchExternalEditorAfterDraw,
 }
