+++ codex-rs/tui/src/lib.rs
@@ -44,9 +44,9 @@ mod clipboard_paste;
 mod color;
 pub mod custom_terminal;
 mod diff_render;
-mod editor;
 mod exec_cell;
 mod exec_command;
+mod external_editor;
 mod file_search;
 mod frames;
 mod get_git_diff;
